[RPC] Ensure _wait_all_workers doesn't swallow exception.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/61094

`_wait_all_workers` was swallowing exceptions and as a result if there
were any errors it would still continue with rpc_agent.join() which would hang
since something already failed before.

To fix this, I've ensured that wait_all_workers throws and in that case we just
proceed with an ungraceful shutdown without joining.
ghstack-source-id: 132902962

Differential Revision: [D29509286](https://our.internmc.facebook.com/intern/diff/D29509286/)